Fix linting and tests
