+++ src/extension.ts
@@ -1,10 +1,10 @@
-// The module 'vscode' contains the VS Code extensibility API
+// The module "vscode" contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-import * as vscode from 'vscode';
-import * as resolver from './resolver';
-import * as fs from 'fs';
-import * as path from 'path';
-import * as mkdirp from 'mkdirp';
+import * as vscode from "vscode";
+import * as resolver from "./resolver";
+import * as fs from "fs";
+import * as path from "path";
+import * as mkdirp from "mkdirp";
 
 function openFile(fileName: string) {
 	vscode.workspace
@@ -15,7 +15,8 @@ function openFile(fileName: string) {
 function prompt(fileName: string, cb: any) {
 	let options = {
 		placeHolder: `Create ${fileName}?`
-	}
+	};
+
 	vscode.window.showQuickPick(["Yes", "No"], options)
 		.then(function (answer) {
 			if (answer === "Yes") {
@@ -40,19 +41,19 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "rails-go-to-spec" is now active!');
+	console.log("Congratulations, your extension 'rails-go-to-spec' is now active!");
 
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-	let disposable = vscode.commands.registerCommand('rails-go-to-spec.railsGoToSpec', () => {
+	let disposable = vscode.commands.registerCommand("rails-go-to-spec.railsGoToSpec", () => {
 		// Display a message box to the user
 		var editor = vscode.window.activeTextEditor;
 		if (!editor) {
 			return; // No open text editor
 		}
 
-		let document = editor.document;
+		let document: vscode.TextDocument = editor.document;
 		let fileName: string = document.fileName;
 		let related: string = resolver.getRelated(fileName);
 		let fileExists: boolean = fs.existsSync(related);
@@ -63,7 +64,7 @@ export function activate(context: vscode.ExtensionContext) {
 			related = resolver.getControllersRelated(fileName);
 			fs.existsSync(related) ? openFile(related) : openPrompt(related);
 		} else {
-			openPrompt(related)
+			openPrompt(related);
 		}
 	});
